(refactor) separate selectTrip to its own component
